restructured to pure Vulkan usage (no more wrapper needed
