+++ Aurora/Source/Renderer/Shader.cpp
@@ -1,4 +1,4 @@
-#include "Renderer/VulkanSpecifics/VulkanShader.h"
+#include "Renderer/Shader.h"
 
 #include <shaderc/shaderc.h>
 #include <spirv_reflect.h>
