+++ Aurora/Source/Renderer/VMA.cpp
@@ -1,4 +1,4 @@
-#include "Renderer/VulkanSpecifics/VMA.h"
+#include "Renderer/VMA.h"
 
 #define VMA_IMPLEMENTATION
 #include <VMA/vk_mem_alloc.h>
