+++ Aurora/Source/Renderer/VulkanCore.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "Renderer/VulkanSpecifics/VMA.h"
+#include "Renderer/VMA.h"
 #include <vulkan/vulkan.h> // technically not necessary, because VMA implements vulkan.h
-#include "Renderer/VulkanSpecifics/VulkanDebug.h"
+#include "Renderer/VulkanDebug.h"
