+++ Aurora/Source/Renderer/VulkanHelper.cpp
@@ -1,4 +1,5 @@
-#include "Renderer/VulkanSpecifics/VulkanHelper.h"
+#include "Renderer/VulkanHelper.h"
+
 
 #include <vector>
 #include <algorithm>
