+++ Aurora/include/Aurora/ApplicationSettings.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+
+namespace Aurora
+{
+	struct ApplicationSettings
+	{
+		std::string RootPath = "";
+		std::string AssetPath = "";
+
+		ApplicationSettings() {};
+	};
+}
