--- Aurora/include/Aurora/Renderer/APIType.h
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace Aurora {
-	enum class APIType
-	{
-		API_TYPE_NONE = 0,
-		API_TYPE_VULKAN
-	};
-}
