+++ StarFire/src/sfpch.h
@@ -8,6 +8,7 @@
 
 #include <optional>
 
+#include <filesystem>
 #include <chrono>
 
 #include <string>
